refactored common schemas
