+++ server/graphql/schema/common/enums.ts
@@ -1,0 +1,27 @@
+import { gql } from "apollo-server-core";
+
+export default gql`
+  """
+  Choose the website that you want to save for this link
+  """
+  enum nameEnum {
+    GITHUB
+    LINKEDIN
+    INSTAGRAM
+    YOUTUBE
+    TWITTER
+    DEWORK
+    LENS
+    OTHER
+  }
+
+  enum OrderField {
+    PUBLISHED_AT
+    UPDATED_AT
+  }
+
+  enum OrderDirection {
+    ASC
+    DESC
+  }
+`;
