+++ server/graphql/schema/common/index.ts
@@ -1,0 +1,5 @@
+import enums from "./enums";
+import inputs from "./inputs";
+import types from "./types";
+
+export default [enums, types, inputs];
